Extracted Pass2FdrSidecar from MergeNodeTask.Run (Phase A)
* Moved the 2nd-pass FDR sidecar step (reload reconciled features ->
  run 2nd-pass Percolator -> write .2nd-pass sidecars -> reload onto
  stubs) verbatim into a new Pass2FdrSidecar collaborator.
* Reduced MergeNodeTask.Run from 374 LOC to a ~45-line sequencer; the
  parity-locked FirstJoinTask.RunPercolatorFdr core is invoked whole.
* Pure code motion: byte-identical blib on Stellar (modes 1 and 2).

See ai/todos/active/TODO-20260616_ospreysharp_debt_paydown_pr4.md
